Add new TODOs and make TODO list nicer
Adds Sticky Card, GT Patterns, and Beam Former to list. Also modifyed the list
to use TODOs to track progress. Also added past addition to allow list to work
as a pseudo feature list.

changes. Lines starting # with '#' will be ignored, and an empty message aborts
the commit.  # # On branch main # Your branch is up to date with 'origin/main'.