change to correct mixin refmap file??, hopefully fixes #10
No clue why this needed to be changed, but it wouldn't load at all without this
changed so ¯\_(ツ)_/¯. For posterity, it would load fine before the change in
dev, but not when built into a jar and put into Prism. I think this is because
it didn't have the refmap in the jar, but no clue why it just stopped working.