The Great Relicensing
Relicensing under LGPL for copyright reasons. Supposedly this lets it work with Minecraft's propreitaryness, but I'm not too convinced. Anyways GPL *did* prevent some other mods like Botania from being used with it, so here we are.

Additionally relicensing resources to be explicitly under CC BY-NC-SA 4.0. I think this is a good license and it's simpler and clearer-cut compared to using GPL on assets.