Implements Pattern Bus (No GUI)
A pattern bus is intended to be like a crafting card without any of the overhead
of crafting jobs. It pushes the entire pattern atomically, either all of it or
none at all. Additionally it can go through pattern p2ps + has a return
inventory like a provider to make it easier to use.