Refactor Multi P2P Tunnels to be distinct objects from their parts
This splits tunnels into the tunnel itself, the parts, and the logic.

The tunnel is an overarching object that owns the logics and can optionally tick
for tunnel-wide logic.

The logic is an object that's like the parts representation in the tunnel. I'm not
sure why it needs to exist, but it handles the logic related to the tunnel with
the part simply forwarding stuff onto the logic.

The part is the object that's just the physical part. It stores things that need
to be stored per part, and forwards on to the logic as needed.