Improve EU MultiP2P performance when there are many idle outputs.
Do this by keeping a per-tick cache of the output p2ps, so that we can
remove p2ps (for the rest of the tick) once they fail to accept
energy.

Assuming that a machine will accept its full amount of energy in a
single acceptEnergyFromNetwork call, we know that either the call
accepted all of our energy (toSend goes to 0), or it has taken all of
its energy for the tick (and we can skip it for the rest of the tick).