Change Multi P2P Logic to be stored in an Optional
Previously multi p2ps would both directly access their logics and wouldn't have
them be stored in Optionals despite null being an intended and common value to
be there.

This commit first makes logic a private field to hide it, instead sub-classes
must access it via getLogic(). Additionally, the logic is now wrapped in an
Optional. This makes it more explicit that logic might be null and should be
handled. This also makes it easier to do some things, ie run a method on logic
only when it's present.

A side effect of this commit is fixing redstone multi p2ps crashing in some
crashes since it's now impossible for logic to be accessed when it's null.